Move educationEntry into cv folder and wrap some tests around it
